Added geometry customization.
Fixed Tactics SPR loading.

Fixed FODLG.MSG filling.

git-svn-id: https://cvet.by:8443/svn/FOnlineSource@174 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 0ca32283d8bc096e72bdbbb49546fa004bf61538

+++ Source/Critter.h
@@ -311,7 +311,7 @@ public:
 	// Send
 	volatile int DisableSend;
 	bool IsSendDisabled(){return DisableSend>0;}
-	void Send_Move(Critter* from_cr, WORD move_params);
+	void Send_Move(Critter* from_cr, DWORD move_params);
 	void Send_Dir(Critter* from_cr);
 	void Send_AddCritter(Critter* cr);
 	void Send_RemoveCritter(Critter* cr);
@@ -360,7 +360,7 @@ public:
 	void Send_CritterLexems(Critter* cr);
 
 	// Send all
-	void SendA_Move(WORD move_params);
+	void SendA_Move(DWORD move_params);
 	void SendA_XY();
 	void SendA_Action(int action, int action_ext, Item* item);
 	void SendAA_Action(int action, int action_ext, Item* item);
@@ -550,7 +550,7 @@ public:
 	void PingOk(DWORD next_ping){pingOk=true; pingNextTick=Timer::FastTick()+next_ping;}
 
 	// Sends
-	void Send_Move(Critter* from_cr, WORD move_params);
+	void Send_Move(Critter* from_cr, DWORD move_params);
 	void Send_Dir(Critter* from_cr);
 	void Send_AddCritter(Critter* cr);
 	void Send_RemoveCritter(Critter* cr);
